requirements.txt
bs4
requests
glob
pillow
